# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Social Distribution - Ctrl + Alt + Defeat API",
      "description": "API for our backend"
    },
    "paths": {
      "/api/authors/": {
        "get": {
          "summary": "Get all authors (REMOTE)",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "type": "integer",
              "description": "Page number",
              "required": false
            },
            {
              "name": "size",
              "in": "query",
              "type": "integer",
              "description": "Number of items per page",
              "required": false
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "authors"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "author"
                        },
                        "id": {
                          "type": "string",
                          "example": "http://127.0.0.1:5454/authors/1d698d25ff008f7538453c120f581471"
                        },
                        "url": {
                          "type": "string",
                          "example": "http://127.0.0.1:5454/authors/1d698d25ff008f7538453c120f581471"
                        },
                        "host": {
                          "type": "string",
                          "example": "http://127.0.0.1:5454/"
                        },
                        "displayName": {
                          "type": "string",
                          "example": "Greg Johnson"
                        },
                        "github": {
                          "type": "string",
                          "example": "http://github.com/gjohnson"
                        },
                        "profileImage": {
                          "type": "string",
                          "example": "https://i.imgur.com/k7XVwpB.jpeg"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "tags": ["Author"]
        }
      },
      "/api/authors/{author_id}/": {
        "get": {
          "summary": "Get a specific author (REMOTE)",
          "parameters": [
            {
              "name": "author_id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "author"
                  },
                  "id": {
                    "type": "string",
                    "example": "http://127.0.0.1:5454/authors/9de17f29c12e8f97bcbbd34cc908f1baba40658e"
                  },
                  "host": {
                    "type": "string",
                    "example": "http://127.0.0.1:5454/"
                  },
                  "displayName": {
                    "type": "string",
                    "example": "Lara Croft"
                  },
                  "url": {
                    "type": "string",
                    "example": "http://127.0.0.1:5454/authors/9de17f29c12e8f97bcbbd34cc908f1baba40658e"
                  },
                  "github": {
                    "type": "string",
                    "example": "http://github.com/laracroft"
                  },
                  "profileImage": {
                    "type": "string",
                    "example": "https://i.imgur.com/k7XVwpB.jpeg"
                  }
                }
              }
            }
          },
          "tags": ["Author"]
        }
      },
      "/api/authors/login": {
        "post": {
          "summary": "Login",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object"
              }
            }
          },
          "tags": ["Author"]
        }
      },
      "/api/authors/update_username": {
        "post": {
          "summary": "Update username",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "new_username": {
                    "type": "string"
                  },
                  "authorId": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object"
              }
            }
          },
          "tags": ["Author"]
        }
      },
      "/api/authors/update_password": {
        "post": {
          "summary": "Update password",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "new_password": {
                    "type": "string"
                  },
                  "authorId": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object"
              }
            }
          },
          "tags": ["Author"]
        }
      },
      "/api/authors/{author_id}/liked": {
        "get": {
          "summary": "Get liked posts for a specific author (REMOTE)",
          "parameters": [
            {
              "name": "author_id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "@context": { "type": "string" },
                        "author": {
                          "type": "object",
                          "properties": {
                            "displayName": { "type": "string" },
                            "github": { "type": "string" },
                            "host": { "type": "string" },
                            "id": { "type": "string" },
                            "profileImage": { "type": "string" },
                            "type": { "type": "string" },
                            "url": { "type": "string" }
                          }
                        },
                        "object": { "type": "string" },
                        "summary": { "type": "string" },
                        "type": { "type": "string" }
                      },
                      "required": ["@context", "author", "object", "summary", "type"]
                    }
                  },
                  "type": { "type": "string" }
                },
                "required": ["items", "type"]
              }
            }
          },
          "tags": ["Author"]
        }
      },
      "/api/authors/{author_id}/inbox": {
        "post": {
          "summary": "Send a like to the author_id of the post (REMOTE)",
          "parameters": [
            {
              "name": "author_id",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "like_author_id": {
                    "type": "string"
                  },
                  "post_id": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object"
              }
            }
          },
          "tags": ["Author"]
        }
      },
...
      "/api/authors/github/{author_id}": {
        "get": {
          "summary": "Get Github username of author",
          "parameters": [
            {
              "name": "author_id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object"
              }
            }
          },
          "tags": ["Author"]
        }
      },
      "/api/authors/github": {
        "post": {
          "summary": "Set Github username",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "github": {
                    "type": "string"
                  },
                  "author_id": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object"
              }
            }
          },
          "tags": ["Author"]
        }
      
      },
...
          "/api/posts": {
            "get": {
              "summary": "Get posts",
              "parameters": [
                {
                  "name": "author_id",
                  "in": "path",
                  "type": "string",
                  "description": "ID of the user",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object"
                  }
                }
              },
              "tags": ["Posts"]
            },
            "post": {
              "summary": "Create a new post",
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object"
                  }
                }
              },
              "tags": ["Posts"]
            }
          },
...
    }
  }